Extra fee implementation implemented fully and DataLoader class added (data at the start)

+++ src/main/java/ee/fujitsu/delivery/app/entity/extrafee/AirTemperatureFeeEntity.java
@@ -19,8 +19,8 @@ public class AirTemperatureFeeEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private Integer lowerBound;
-    private Integer upperBound;
+    private Float lowerBound;
+    private Float upperBound;
     private Long vehicleId;
     private BigDecimal fee;
 }
